scripts/ci/requirements.txt
PyYAML==6.0.2
gitlint==0.19.1
